requirements.txt
# ShellIA - Requirements

# Framework web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Providers IA
openai>=1.12.0          # Pour ChatGPT
anthropic>=0.39.0       # Pour Claude

# Utilitaires
pydantic>=2.6.0         # Validation de données (inclus avec FastAPI)
python-dotenv>=1.0.0    # Gestion des variables d'environnement via .env
paramiko>=3.4.0         # Connexion SSH pour exécution distante
pexpect>=4.9.0          # Sessions shell persistantes (local et WSL)

# Authentification et sécurité
python-jose[cryptography]>=3.3.0  # JWT tokens
bcrypt>=4.0.0                     # Hachage de mots de passe
python-multipart>=0.0.9           # Pour les formulaires de login
google-auth>=2.27.0               # Google OAuth2
google-auth-oauthlib>=1.2.0       # Google OAuth2 helpers
google-auth-httplib2>=0.2.0       # Google Auth HTTP
msal>=1.28.0                      # Microsoft OAuth2 (Hotmail/Outlook/Azure AD)
httpx>=0.27.0                     # HTTP client pour Facebook OAuth2

# Optionnel : Pour le développement
# pytest>=8.0.0         # Tests unitaires
# black>=24.0.0         # Formatage de code
# ruff>=0.2.0           # Linter rapide
